Add availability for the baby sitters added.

+++ booking/helper_methods.py
@@ -1,4 +1,6 @@
 from booking.models import *
+import pytz
+utc=pytz.UTC
 
 
 # Finds overlapping intervals
@@ -16,11 +18,10 @@ def overlap(times, interval):
 # Merges intervals and returns the merged interval
 def merge_interval(times, interval):
     overlapped_intervals = overlap(times, interval)
-    print(overlapped_intervals)
     merged_interval = interval
     for i in overlapped_intervals:
         merged_interval.start_time = min(times[i].start_time, merged_interval.start_time)
-        merged_interval[1] = max(times[i].end_time, merged_interval.end_time)
+        merged_interval.end_time = max(times[i].end_time, merged_interval.end_time)
     return merged_interval
 
 
@@ -31,16 +32,25 @@ def included(times, interval):
     return -1
 
 
-# Input are a baby sitter object and an availability object
+# Input is an availability object
 # It finds all the overlapping availabilities and deletes them and adds the new availability to database
-def add_availability(baby_sitter, availability):
-    booked_times = BookedTime.objects.filter(baby_sitter=baby_sitter)
+def add_availability(availability):
+    availability.start_time = utc.localize(availability.start_time)
+    availability.end_time = utc.localize(availability.end_time)
+    booked_times = BookedTime.objects.filter(baby_sitter=availability.baby_sitter)
     if len(overlap(booked_times, availability)) > 0:
         print("This time cannot be available because it's booked.")
-        return
-    availabilities = Availability.objects.filter(baby_sitter=baby_sitter)
+        return "This time cannot be available because it's booked."
+    if availability.start_time >= availability.end_time:
+        print("Start time must be sooner than end time")
+        return "Start time must be sooner than end time"
+    print(availability.baby_sitter)
+    availabilities = Availability.objects.filter(baby_sitter=availability.baby_sitter)
     overlapped_availabilities = overlap(availabilities, availability)
-    new_availability = merge_interval(availability)
+    new_availability = merge_interval(availabilities, availability)
+    print(new_availability.start_time)
+    print(new_availability.end_time)
     for i in overlapped_availabilities:
         availabilities[i].delete()
     new_availability.save()
+    return "availability successfully added."
